refactor(exex): move notifications to a separate module
