+++ crates/exex/exex/src/lib.rs
@@ -46,6 +46,9 @@ pub use event::*;
 mod manager;
 pub use manager::*;
 
+mod notifications;
+pub use notifications::*;
+
 mod wal;
 pub use wal::*;
